Backend/requirements.txt
Flask==3.0.3
FastAPI==0.115.0
SQLAlchemy==2.0.36
APScheduler==3.10.4
Flask-Mail==0.9.1
python-dotenv==1.0.1